refactor: address review - authenticate naming, per-platform files, stable assembly inputs
Rename the authenticity phase to AuthenticateQuoteV3 (validate remains the comparison phase); split the SEV and TDX authenticators into their own files; group the envelope and code-provenance expectations into ExpectedValues so the assembly signature stays stable; construct TDX verify options explicitly; simplify ReferenceValuesCollateral to a value/error pair with ErrCollateralNotFound.
